refactoring file structures
